+++ demos/chinook/src/main/java/is/codion/demos/chinook/ui/InvoiceConditionPanel.java
@@ -190,8 +190,8 @@ final class InvoiceConditionPanel extends TableConditionPanel<Attribute<?>> {
 		}
 
 		private void activate() {
-			customerConditionPanel.condition().operator().set(Operator.IN);
-			dateConditionPanel.condition().operator().set(Operator.BETWEEN);
+			customerConditionPanel.model().operator().set(Operator.IN);
+			dateConditionPanel.model().operator().set(Operator.BETWEEN);
 			customerConditionPanel.requestInputFocus();
 		}
 
@@ -254,7 +254,7 @@ final class InvoiceConditionPanel extends TableConditionPanel<Attribute<?>> {
 
 			private DateConditionPanel(ConditionModel<LocalDate> conditionModel) {
 				super(conditionModel);
-				condition().operator().set(Operator.BETWEEN);
+				model().operator().set(Operator.BETWEEN);
 				updateCondition();
 				initializeUI();
 			}
@@ -297,8 +297,8 @@ final class InvoiceConditionPanel extends TableConditionPanel<Attribute<?>> {
 			}
 
 			private void updateCondition() {
-				condition().operands().lower().set(lower());
-				condition().operands().upper().set(upper());
+				model().operands().lower().set(lower());
+				model().operands().upper().set(upper());
 			}
 
 			private LocalDate lower() {